backend-python/pyproject.toml
[project]
name = "sky-backend"
version = "0.1.0"
description = "Sky Finance — backend Python (FastAPI + ARQ workers)"
requires-python = ">=3.12"
license = { text = "UNLICENSED" }
authors = [
    { name = "Cristian Vásquez Guevara" },
    { name = "Juan José Latorre Pérez" },
]

dependencies = [
    # API
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",

    # Database
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "supabase>=2.10.0",

    # Queue / Workers
    "arq>=0.26.1",
    "redis>=5.2.0",

    # Scraping
    "playwright>=1.49.0",

    # AI
    "anthropic>=0.40.0",

    # Crypto — AES-256-GCM compatible binario con Node.js
    "cryptography>=44.0.0",

    # HTTP client (Fintoc, webhooks, APIs bancarias)
    "httpx>=0.28.0",

    # Observability
    "structlog>=24.4.0",
    "prometheus-client>=0.21.0",

    # Auth
    "PyJWT>=2.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "httpx",  # TestClient uses httpx
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "S", "B", "SIM", "RUF"]
ignore = ["S101"]  # allow assert in tests

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
